refactor(lowy+hickey): move TerminalTag → intent/IntentGlyph, hoist Segmenter
Lowy: the chip-tier renderer of intent is a sibling to IntentMarkdownBlock
(block tier) and the canonical intent surface lives in packages/client/src/
intent/. Co-locate all three rendering tiers in the intent module.
Hickey cross-val: the move alone leaves a 'TerminalTag in intent/'
naming mismatch — rename to IntentGlyph at the same time.

Hickey D: while there, hoist the stateless Intl.Segmenter out of
firstGrapheme() so it isn't reallocated on every reactive update.
